+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Upgrade presets for TensorFlow 1.5.0
  * Include `facemark.hpp`, `facemarkLBF.hpp`, `facemarkAAM.hpp`, `face_alignment.hpp` from the `opencv_face` module
  * Add `AxpyLayer` to presets for Caffe ([pull #508](https://github.com/bytedeco/javacpp-presets/pull/508))